iseq.h (struct iseq_catch_table_entry): shrink to 32 bytes on 64-bit
* iseq.h (struct iseq_catch_table_entry): shrink to 32 bytes on 64-bit

This drops a few minor page faults on my system at startup.
Using 64-bit start, end, cont, and sp fields are unnecessary as
label_get_position and label_get_sp only return int values.
label_get_position and label_get_sp should probably return
unsigned, but that is a future change.

git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@46813 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ iseq.h
@@ -62,10 +62,10 @@ struct iseq_catch_table_entry {
 	CATCH_TYPE_NEXT   = INT2FIX(6)
     } type;
     VALUE iseq;
-    unsigned long start;
-    unsigned long end;
-    unsigned long cont;
-    unsigned long sp;
+    unsigned int start;
+    unsigned int end;
+    unsigned int cont;
+    unsigned int sp;
 };
 
 PACKED_STRUCT_UNALIGNED(struct iseq_catch_table {
